DevOps/Docker/Лекции_нетологии/Docker/Project-shop/requirements.txt
appnope==0.1.2
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
coverage==5.5
decorator==5.0.9
Django==3.2
django-debug-toolbar==3.2.1
django-emoji==2.2.2
django-filter==2.4.0
djangorestframework==3.12.4
emoji==1.2.0
Faker==8.10.3
Flask==2.0.0
greenlet==1.1.1
idna==2.10
iniconfig==1.1.1
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
mimesis==4.1.3
model-bakery==1.3.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.0
xmltodict==0.12.0
